requirements.txt
pip
autopep8
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.6
python-dotenv
google-generativeai
requests
groq
Flask-SQLAlchemy
Flask-Migrate
